requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Google API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# HTTP requests
requests>=2.31.0

# Web scraping (optional, add as needed)
# beautifulsoup4>=4.12.0
# selenium>=4.15.0

# AI APIs (optional, uncomment as needed)
# openai>=1.3.0
# anthropic>=0.7.0
